requirements.txt
# Core dependencies
openai>=1.0.0
pinecone-client>=3.0.0
python-dotenv>=1.0.0

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Utilities
requests>=2.31.0
python-dateutil>=2.8.0
